% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ismb12-data.R
\docType{data}
\name{ismb12}
\alias{ismb12}
\title{Twiter data for 2012}
\format{An object of class \code{data.frame} with 3162 rows and 5 columns.}
\usage{
load("./data/ismb2012.RData")
}
\description{
Twiter Data to generate ranking
}
\keyword{datasets}

