% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ismb14-data.R
\docType{data}
\name{ismb14}
\alias{ismb14}
\title{Twiter data for 2014}
\format{An object of class \code{data.frame} with 3235 rows and 5 columns.}
\usage{
ismb14
}
\description{
Twiter Data to generate ranking
}
\keyword{datasets}

