% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ismbTweetAnalysis-package.R
\docType{data}
\name{ismb2014}
\alias{ismb2014}
\title{ISMB tweets from 2014}
\format{A data frame with 3235 rows and 5 variables}
\source{
\url{https://github.com/stephenturner/twitterchive}
}
\description{
A set of tweets from ISMB 2014 meeting collected by Stephen Turner
}
\details{
\itemize{
  \item text. the text of the tweet
  \item created. the date and time of the tweet
  \item id. the twitter id of the tweet
  \item screenName. who sent the tweet
  \item hashSearch. the hashtag that was searched
}
}

