% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ismbTweetAnalysis-package.R
\name{totalRT}
\alias{totalRT}
\title{total retweeted}
\usage{
totalRT(tweetDF, countColumn)
}
\arguments{
\item{tweetDF}{data.frame of tweets}

\item{countColumn}{which column has the retweet counts (should be a text id)}
}
\value{
data.frame of counts
}
\description{
how often a users tweets were retweeted
}

