% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ismbTweetAnalysis-package.R
\name{tweetCounts}
\alias{tweetCounts}
\title{user tweet counts}
\usage{
tweetCounts(tweetDF)
}
\arguments{
\item{tweetDF}{data frame of tweets}
}
\value{
data.frame with user, total, original, retweet
}
\description{
calculates the number of tweets for each user, giving the total tweets, 
total that were retweets, and total that were original. Assumes that a retweet
starts with the text "RT"
}

