% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ismbTweetAnalysis-package.R
\name{tweetRank}
\alias{tweetRank}
\title{total tweet rank}
\usage{
tweetRank(tweetCounts)
}
\arguments{
\item{tweetCounts}{a vector of tweets by user}
}
\value{
ranks by percentage of total from tweetCounts
}
\description{
calculates rank of a user by the percentage of total tweets
}

